break out nav item into component
